# speed up parsing