## Context

When testing large ansi files I found that they can take a while to load. After reading ansiparser, the main culprit of the slow parsing is the use of the regex to detect alpha characters. The python standard lib has a `str.isalpha` method that works well and is much faster.

I did a quick demo in ipython

```python
In [1]: import re

In [2]: s = 'A'

In [3]: %timeit re.search('[a-zA-Z]', s)
316 ns ± 12 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [4]: %timeit s.isalpha()
11.8 ns ± 0.0565 ns per loop (mean ± std. dev. of 7 runs, 100,000,000 loops each)
```

## Changes

- Convert the regex search into a for loop, calling `isalpha()` on each char

## Effect

On my PC, this reduces the time to load `Blocktronics-WTF4_Megajoint.ans` from **44 seconds > 2.5 seconds**

As far as I can see, the behaviour between this version and the existing one is identical and doesn't produce any different results that I've seen so far.